requirements.txt
autopep8==1.5.7
Django==3.2.3
django-cors-headers==3.7.0
django-debug-toolbar==3.2.1
djangorestframework==3.12.4
pre-commit==2.13.0
pylint==2.8.2
pytest==6.2.4
python-decouple==3.4
